+++ engine/src/main/java/net/jqwik/engine/descriptor/DiscoverySupport.java
@@ -1,8 +1,10 @@
 package net.jqwik.engine.descriptor;
 
+import java.lang.annotation.*;
 import java.lang.reflect.*;
 import java.util.*;
 import java.util.function.*;
+import java.util.logging.*;
 
 import org.junit.platform.engine.*;
 
@@ -12,6 +14,9 @@ import static java.util.stream.Collectors.*;
 import static org.junit.platform.commons.support.AnnotationSupport.*;
 
 public class DiscoverySupport {
+
+	private static final Logger LOG = Logger.getLogger(DiscoverySupport.class.getName());
+
 	public static Set<TestTag> findTestTags(AnnotatedElement element) {
 		return findRepeatableAnnotations(element, Tag.class)
 			.stream()
@@ -38,5 +43,15 @@ public class DiscoverySupport {
 		return () -> nameSupplier
 						 .get()
 						 .replaceAll("_", " ");
+	}
+
+	public static void warnWhenJunitAnnotationsArePresent(AnnotatedElement element) {
+		Annotation[] directAnnotations = element.getDeclaredAnnotations();
+		for (Annotation annotation : directAnnotations) {
+			if (annotation.annotationType().getPackage().getName().startsWith("org.junit")) {
+				String message = String.format("[%s] has annotation [%s] from JUnit which cannot be processed by jqwik", element, annotation);
+				LOG.warning(message);
+			}
+		}
 	}
 }